perf(client): use a single file for `HttpClient`
The `HttpClient` class used to inherit `HttpClientBase`, a base class
initially included in case the scope of this project changes.

Up until now the scope of this project didn't expand, thus
removing the need to spread functionality across multiple files.

Due to the above changes, the build output decreases by ~25kb.
