+++ src/client/http-client.ts
@@ -16,20 +16,14 @@
  */
 
 import {
+  AxiosRequestConfig,
   delete as makeDeleteRequest,
   get as makeGetRequest,
   patch as makePatchRequest,
   post as makePostRequest,
 } from 'axios';
 
-import { HttpClientBase, HttpClientBaseOptions, HttpRequestConfigInit } from './http-client-base';
-
-/**
- * Options to customize how a Gelato {@link HttpClient} processes requests.
- *
- * @publicApi
- */
-export type HttpClientOptions = HttpClientBaseOptions;
+import { GELATO_API_HEADER_KEY } from '../utils/env';
 
 /**
  * HTTP client for making (authenticated) requests to Gelato APIs.
@@ -50,7 +44,9 @@ export type HttpClientOptions = HttpClientBaseOptions;
  *
  * @publicApi
  */
-export class HttpClient extends HttpClientBase {
+export class HttpClient {
+  constructor(private options_: HttpClientOptions) {}
+
   /**
    * Make a `GET` request to an eligable Gelato API endpoint.
    * @param url Resource URL to fetch.
@@ -58,9 +54,8 @@ export class HttpClient extends HttpClientBase {
    * @returns
    */
   async get<T>(url: string, options: HttpRequestConfigInit = {}): Promise<T> {
-    const config = this.getRequestConfig(options);
+    const config = prepareRequestConfig(this.options_.apiKey, options);
     const response = await makeGetRequest<T>(url, config);
-
     return response.data;
   }
 
@@ -70,7 +65,7 @@ export class HttpClient extends HttpClientBase {
    * @param options Options to customize the request.
    */
   async post<T, D>(url: string, data: D, options: HttpRequestConfigInit = {}): Promise<T> {
-    const config = this.getRequestConfig(options);
+    const config = prepareRequestConfig(this.options_.apiKey, options);
     const response = await makePostRequest<T>(url, data, config);
     return response.data;
   }
@@ -80,7 +75,7 @@ export class HttpClient extends HttpClientBase {
    * @param options Options to customize the request.
    */
   async patch<T, D>(url: string, data: D, options: HttpRequestConfigInit = {}) {
-    const config = this.getRequestConfig(options);
+    const config = prepareRequestConfig(this.options_.apiKey, options);
     const response = await makePatchRequest<T>(url, data, config);
     return response.data;
   }
@@ -90,8 +85,43 @@ export class HttpClient extends HttpClientBase {
    * @param options Options to customize the request.
    */
   async delete<T = unknown>(url: string, options: HttpRequestConfigInit = {}) {
-    const config = this.getRequestConfig(options);
+    const config = prepareRequestConfig(this.options_.apiKey, options);
     const response = await makeDeleteRequest<T>(url, config);
     return response.data;
   }
+}
+
+export type HttpRequestHeadersInit = Omit<
+  AxiosRequestConfig['headers'],
+  'Content-Type' | 'content-type' | 'X-API-KEY'
+>;
+
+export type HttpRequestConfigInit = Omit<AxiosRequestConfig, 'headers' | 'data'> & {
+  headers?: HttpRequestHeadersInit;
+};
+
+/**
+ * Options to customize how a Gelato {@link HttpClient} processes requests.
+ */
+export interface HttpClientOptions {
+  apiKey?: string;
+}
+
+/**
+ * Merge an *incoming* (axios) request config object with internal http config.
+ * @param requestOptions - Incoming request config object.
+ * @returns The merged Axios request config object.
+ */
+function prepareRequestConfig(apiKey: string | undefined, reqConfig: HttpRequestConfigInit) {
+  const { headers: headers_, ...otherOptions } = reqConfig;
+  const headers: Record<string, string> = {
+    ...headers_,
+    'Content-Type': 'application/json',
+  };
+
+  if (apiKey) {
+    headers[GELATO_API_HEADER_KEY] = apiKey;
+  }
+
+  return { ...otherOptions, headers };
 }
